Use `converse-avatar` for roster items
