+++ src/shared/styles/_core.scss
@@ -396,12 +396,6 @@
         width: 100%;
     }
 
-    .avatar {
-        border-radius: var(--avatar-border-radius);
-        border: var(--avatar-border);
-        background-color: var(--avatar-background-color);
-    }
-
     .avatar-autocomplete {
         margin-right: 0.5em;
         vertical-align: middle;
